Moved Sound class to different file. Fixed bug where song stops when new game begins. Fixed bug where amount bet wasn't passed to gui
